source/vrf_request_lambda/pyproject.toml
[tool.poetry]
name = "create_vrf_request_lambda"
version = "0.1.0"
description = ""
authors = ["Eugene"]

[tool.poetry.dependencies]
python = "^3.9"
beartype = "^0.10.4"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
aws-lambda-powertools = "^1.28.0"
boto3 = "^1.24.61"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
